Add measurement from nav start to just before 'onload' event
This is apparently the base figure that GA computes for 'page load time'. It is
not very granular and thus less helpful than some of the other measurements,
but until we ramp up our capacity to analyze the data we can use this figure as
an overall indicator of user-perceived latency.

Change-Id: Ide5f832e1e88ef998ad010efb77c4360c0333492