Work around noncompliant IE9 behavior
The specs from the W3C specify that "if there is no previous document, this
attribute [i.e., navigationStart] must return the same value as fetchStart."
IE9 incorrectly sets the value to 0 when there is no previous document. As a
result, the values for 'loading' and 'sending' are left as epoch timestamps
(rather than time deltas). These overflow INT in MySQL, leaving us with a bunch
of 2147483647 values (equal to max. int).

Solution: explicitly check if navigationStart is 0; use fetchStart if it is.

Bug: 46474
Change-Id: I31ddc1b26923891899cb2e7614f4040fc0d36ecf