Update schema to 5832704
The version of NavigationTiming that is currently deployed computes a set of
Navigation Timing intervals that represent a small subset of the intervals that
are potentially interesting to us. A more flexible approach is to simply
log the value of each interesting marker relative to NavigationStart, and
derive interval measurements on the server.

This patch updates NavigationTiming to revision 5832704 of
<https://meta.wikimedia.org/wiki/Schema:NavigationTiming>. The data points that
are collected reflect my experiments with the Navigation Timing API, documented
at <https://wikitech.wikimedia.org/wiki/Navigation_Timing>.

Change-Id: Ib43387e72ea309f547d5cd4fc52f24cead0e6305