Don't report stats when page visibility changes during page load
Browsers sometime defer or deprioritize painting or executing code on
pages that load in the background (for example, in an inactive browser
tab). We don't want to report metrics in these cases, because a slow
page load is not necessarily a bad thing if it is invisible to the user
and if the slowdown is in service of making some operation that is
user-facing quicker.

Bug: T146510
Change-Id: I8e8ec96fc99f1e38d254f421c5f48cb313ba36e5