ext.NavigationTiming: Restore unsampled Save Timing
Follows-up 8957895a6, which wrongly assumed that Save Timing
should be in the same sampling condition as Navigation Timing.

Restore the same logic as prior to 8957895a6:

 onLoadComplete:
 -> inSample -> do Navigation Timing
 -> Save Timing

With the only difference that we still compute inSample before
onLoadComplete happens, so that we can start preloading the
schema modules earlier.

Inside onLoadComplete() simply give Save Timing its own using()
call instead of re-using the same variable.

RL will still naturally de-duplicate and re-use the existing
deferred internally (if not resolved already).

If we're in the sampling, load both modules to avoid having
to potentially make 2 requests instead of 1.
On postEdit, we'll either no-op if the module is already loaded,
or we'll on-demand load just Save Timing if needed.

Bug: T161368
Change-Id: I45583feaa33936f129ca96a56341463faed8b2a8