Unbreak navTiming reporting
Follows-up 3d5946f. The check for '> 0' of the computed relative value
being removed in itself was fine, but the problem was that this check
was also indirectly used to ensure the original value exists at all
and is in fact not also 0.

We could add a check for the original value (before subtraction)
is above 0

For example: The raw 'secureConnectionStart' value (expected to be
a timestamp) may be 0 itself, in which case subtracting navStart
produces a negative value. Also, if the key is absent, it would produce
NaN which (in JSON) becomes null, which is invalid per the schema
declaring it as int, thus the entire beacon packet was being discarded.

Bug: T176105
Change-Id: I187505b75ad8f94763d20546eb408139622c617a