ext.NavigationTiming: Allow oversampling based on geography or user agent
This code allows oversampling based on geography (specifically, country) or user agent.  Configuration is controlled by
a wg parameter.  The result is a standard NavTiming event that goes to the same topic as any other, but with a new
parameter "isOversample" that is set to true.  We'll need to make appropriate changes to the script that processes
NavTiming events in order to ensure that these events do not pollute the standard perf timing population.

Backend changes will be made independently.

Relies on changes for T182262 (already rebased off of that changeset)

Bug: T181413
Change-Id: I0908bad313a8566cd0a47e85092bf8cb3a3a9204