Add mediaWikiLoadEnd, remove mediaWikiLoadComplete
Per discussion at T160315, the mediaWikiLoadStart offset and
mediaWikiLoadComplete delta metrics are not useful.

We should track mediaWikiLoadEnd instead.

Schema diff at https://meta.wikimedia.org/w/index.php?diff=17957960&oldid=17854363

Server-side processing is already in place because we've had this
metric before, in 2015, as part of an experiment that added a lot
of metrics, but had to revert because the beacon became too big.
This isn't an issue now because we're only adding 1, and even
removing another.

Bug: T160315
Change-Id: Ibff59c4bc6e6467efc48b3453ec52a776be129d1