tests: Simplify the event object assertions
Compare either the type (if value is variable or foreign), or
strictly compare the value (which already checks the type).
Doing both was getting a bit tedious and not actually useful.

It already had this concept (by determining whether the array
description has a [1] value index), but it should be cleaner
now.

Change-Id: I9dc45d627a52294b7a041d8d2437a415e2b0fa31