ext.navigationTiming: Don't send beacon without basic NavTiming support
The only metric we currently send in cases where we are in-sample
and there is no NavTiming API in the browser, is mediaWikiLoadEnd.
(Formerly, mediaWikiLoadComplete, per e227e88f33598).

In its old form mediaWikiLoadComplete was a delta from Start to End,
where Start and End were measured as absolute clock times using
mw.now(), which is basically Date.now() but enhanced to use
perf.timing.navStart+perf.now() in supported browsers.

This delta wasn't useful (per T160315), and we want the 'End'
metric only, but for it to be useful we need it relative to
navigationStart.

In e227e88f33598, I made the mistake of logging the value from
mw.now() which isn't useful because it just goes up in the graph.
It needs to have deducted the timestamp of the start of the
navigation, which depends on having the Navigation Timing API.

In preparation for fixing that metric, remove reporting from
non-NavTiming browsers. Again, none of our metrics are affected
by this because the only metric we were actually reporting
without NavTiming, is mediaWikiLoadEnd itself, which we only
introduced last week, is broken, and requires NavTiming to work.

Bug: T160315
Change-Id: Ifd4da292743d83c899cfcafc5ba8747c8f24c25a