ext.navigationTiming: Fix mediaWikiLoadEnd metric (offset, not timestamp)
It was reporting the full timestamp to EventLogging/Graphite,
which is producing a odd graph essentially plotting current time
each minute, give or take some clock skew and network latencies.

While an interesting excercise in its own, let's fix this to
report an offset like all other metrics, relative to navStart.

The main mistake was that it accidentally used mw.now() instead
of the similarly named performance.now(), where perf.now() starts
at zero each navigation going up monotonically, and mw.now() is
more like clock-time. (mw.now is actually navStart+perf.now()).

Add a general regression test to the 'Basic' test case where all
timing metrics are asserted to be offsets instead of timestamps,
using the same logic we use elsewhere (such as in BagOStuff) by
asserting the value is less than 1 year in seconds (or, from the
perspective of time, that it is less than 1971-01-01), which is
a very liberal way to catch this bug.

Bug: T160315
Change-Id: Ib02518242289c27bd2c58d81956297982e483df1