ext.navigationTiming: Listen for window.onload, not document.onload
Follows-up d50675a, which was listening for 'load' on document instead
of window. As a result, any page views where the code runs before the
load event would not have their beacon sent.

This caused about 30% drop in metric volume, and in terms of numbers,
affected mostly p75 and p99, as expected (in retro-spect).

Many global events in the DOM fire on document and window, except 'load'
which is window-only. This despite readyState and onreadystatechange being
document-only, as well as <body onload>. Oh well, legacy webs :-)

Bug: T193570
Change-Id: If293a156cac7162164eb2682892d725f2f88aff5