Remove 'visibilitychange' handler when it is no longer needed
Early on, we add a listener for the 'visibilitychange' event.
If the event has not fired by the time the page has completed loading,
we are no longer interested in the event and should remove it.

I noticed in Dev Tools timeline that whenever I instrumented
a repeat view or second view, the teardown for the current view
starts by handling the 'visibilitychange' event because we have
a listener for it. Aside from switching contexts, and setting an
unused variable, it also got added overhead from one(), which also
removes the event handler explicitly.

Moving this to loadCallback means in the common case of a normally
viewed page, we'll remove it before it fires with no DOM overhead
besides the adding and removing of the handler.

Bug: T192862
Change-Id: Ib5ee65623b139792e4ae1da39fbc362618c57c68