Use mw.eventLog.inSample() instead of copying it
It originally written here when the inSample() method didn't
exist. Then at some point it started to exist, then it switched
to using crypto instead of Math.random(), and we updated our
code to match. Then I retroactively documented in Ib96f794 with
Ian that we needed our own copy to avoid loading an extra
dependency on EventLogging core.

This was incorrect because the method isn't part of EL core, it's
part of EL's lightweight bootstrap (ext.eventLogging.subscriber)
that already loads on all page views. We corrected the comment
in code review, but didn't revisit the underlying assumption.

Remove our copy and use EL's method instead.

Also, make the dependency on EventLogging explicit in\
extension.json. This is already documented on mediawiki.org,
and our CI layer also is configured to install EventLogging
when testing NavigationTiming. The ability to require an
extension is fairly new, so now that we can do this, do it, which
gives developers a more useful error message if they forget to
install it.

Bug: T192862
Change-Id: I5c6f38b34df2fa594a81f04162d3b37d9b624900