Fix actual Heisenbug – tests fail if you look away
This didn't affect when testing with Headless Chrome via
'npm run qunit', but when running the tests via
Special:JavaScriptTest in your own browser, and tabbed away
or switched windows, they'd fail when you get back, but pass if
you refresh and keep focus on it.

The culprit is that 'visibilityChanged' isn't mocked, so
whatever value it comes to naturally from the loading of NavTiming
on the special page will preserve even if calling loadCallback()
a second time in the test.

Change-Id: Ied148cb1cb0c7d9dba2eda2725c41203a8cfd2db