ext.navigationTiming: Make sure setMwLoadEnd() is not deferred
When ResourceLoader internals (mw.loader#doPropagation) notice
that the mw.loader.using() job has its dependencies satisfied,
the call to setMwLoadEnd() should happen in the same tick as
that realisation.

This was not the case due setMwLoadEnd() being inside onLoadComplete()
instead of onMwLoadEnd(), so it was getting deferred along with the
rest of the async-then() callback. This meant that it was allowing
a potentially large amount of other code to be executed synchronously
(e.g. other modules that arrived in the same batch as our deps),
as well as other async tasks with a similar timeout queued before
our then(), and any rendering/non-JS task the browser decides to
do. Fix this by using done() instead within onMwLoadEnd().

Before/after comparison at
 <https://phabricator.wikimedia.org/T204426#4586149>.

Bug: T204426
Change-Id: I3f8a19480aad26b6b6ce2b9c632282871af1b217