ext.navigationTiming: Improve setMwLoadEnd() tests
The existing test case was calling state() and then making
assertions about whether the callback had happened yet.

This logic was flawed because the callback passed to then()
is itself asynchronous, which means even if the code was broken
and decided to stop waiting after the first failing module, the
test would still pass because it could not have been called yet.

Fix this by mocking async logic.

Also:

* rename 'queue' to 'log' for clarity.

* re-order the state changes for 'fail' to happen first, because
  this is a regression test for our logic waiting for all
  listed modules to settle. It must not give up after the first
  failure.

* add valid return value to one of the 'getEntriesByType' mocks.
  I'm not sure but 1/10 times, locally, the real loadComplete
  (from loading the unit test page) happens to fire right in
  the middle of this test and the real code then invokes
  getFirstPaint() where the try/catch doesn't fail because this
  is an empty function, but then it crashes on
  `if ( paintEntries.length )` with:
  > TypeError: Cannot read property 'length' of undefined
  Follows-up 61d41eab8806.
  In Jenkins this seems to be happening almost every time
  as of this commit.

Change-Id: I6f847a1a00d4658a09c2d395d01a4061b667c029