ext.navigationTiming: Update mwLoadEnd to use RLPAGEMODULES
Instead of building a list of modules currently in 'loading'
state at the time this script executes, use the list of modules
originally requested for this page.

When this metric was first added to core, I took a shortcut
by using getModuleNames().filter(). That was mostly fine at the
time, because the code building the list was in mediawiki.js,
which would execute before anything else.

When we moved it to ext.navigationTiming.js, it changed its
meaning due to builing the list much much later.

The chain of modules loading, and deciding to load yet other
module is potentially infinite and not what this metric is meant
to track, nor what it is currently tracking. Events trigger from
clicks, hovers, scrolls, etc. there can alway be more.

Fix this so that it isn't subject to race conditions that include
semi-random modules lazy-loaded by unrelated modules that happen
to sometimes execute before NavTiming or in the same batch.

On a plain MW install with only EL+NT, the effect can still be
observed because NT's own lazy-loaded modules (for schemas
and rumSpeedIndex) were also ending up in the list of modules
to wait for. Oops!

Add a test that confirms we only wait for RLPAGEMODULES, and not
an unrelated module.

Bug: T204426
Change-Id: I994c3b8f2dee702adda9e5ca923f7527262657f1