Remove use of deprecated 'schema.*' modules
Phase 2 of 3, of "EventLogging light", has been completed and
these modules are now deprecated empty shells.

Phase 3 will also remove the need for ext.eventLogging, leaving
only the synchronous soft-dependency call to mw.track() for the
dispatching of events (handled by ext.eventLogging.subscriber,
which we don't load ourselves).

After phase 3, it will be possible to synchronously reach
navigator.sendBeacon() from mw.track() events with little to no
overhead and no abortable or asynchronous indirection, which is
currently reason for a ~1s delay between loadEvent and sendBeacon
(depends on latency, device capabilities, and radio activation),
during which the page can be closed and thus abort the event
(whereas once sendBeacon is reached, it can survive).

For Navigation Timing in particular, though, that won't yet be
the case as our preload batch is also used for rumSpeedIndex.
We may want to find another way to load that.

Bug: T192862
Bug: T205744
Change-Id: I7dca2b9f13c354c3517ca527b734716bdde695cd