Restore randomization independence to samplings
mw.eventLog.inSample works on a sticky pageview token.

Which means that it can create "sync" effects when
called for different purposes with population sizes that
are multiples of each other.

To avoid this, use mw.eventLog.randomTokenMatch
which will draw a new random number on every call.
This will ensure that the different oversample
population draws are independent from each other,
as well as the fact that the survey's randomness is
independent from the randomness that triggers
NavigationTiming sampling itself.

Bug: T207411
Change-Id: I71a86c88131b3bab9de57870efcc44250732fd53