Use single/lightweight EventLogging module
There are no longer a separate "subscriber" and "core" modules
of EventLogging. Most of the core logic was for debugging and
moved out. The remaining two pieces are now together as one module
"ext.eventLogging" which acts as the subscriber (that is, loaded
on all pages to handle mw.track events).

The deprecation guide at T205744 recommends using mw.track and no
dependency link to EventLogging, but given that we use the sampling
methods, keeping the dependency the same as before.

But given that "ext.eventLogging.subscriber" is now a deprecated
alias to "ext.eventLogging", it means we can remove various
using() calls and brings us closer to resolving T192862.

The integration test for ServerTiming needed a minor update because
ServerTiming now emits before NavigationTiming (call 0 vs call 1).

Bug: T192862
Change-Id: Ib74e7babf226bcd21cba429c9cf81362a3e655c1