build: Set 'root: true' in eslintrc
While for most rules, an explicit "extends": "../" is required to trigger
inheritence in ESLint, it inherits some settings by default within
a project.

The top file within a project needs `root: true` to prevent unintended
interference from unrelated directories higher in the hierarchy on disk
In particular, it disables the confusing error messages from ESLint
in the Atom Editor due to it evaluating mediawiki-core/.eslintrc.json
two directories up from extensions/NavigationTiming, which currently
uses a different ESLint version in its package.json. We'll update
soon but it's generally okay for the versions to differ slightly
as it's meant to be per-repo.

Change-Id: I6cb039b250ec412fbfcaa07d039a382a44d649f7