tests: Improve stability and visibility of emitCpuBenchmark() test
Stability:

The event is emitted by emitCpuBenchmark from `deferred.then`
callback, however what it returns is `deferred`, not the result of
`deferred.then`. Which means the return value awaits only the
completion of the worker, not the completion of the event.

This is presumably why the `setTimeout` hack existed in the test code
previously. Instead of that hack, actually fix the root problem which
is that we should return the value of `deferred.then` instead, and also
making sure that the nested `batteryPromise.then` is properly returned
and chained along as well.

The emitCpuBenchmark() return value is only used by tests, so this
change should not affect anything in the actual behaviour.

Visibility:

Instead of the sinon call assertion, compare the event object
directly. This way, when it fails, it will show what was actually
logged instead of a not-to-useful boolean false.

What is going wrong is that sometimes setupFirstInputTimingObserver is
firing in the middle of this, and so a different even appears
here first. I will deal with that in a follow up commit.

Bug: T299780
Depends-On: Ia8ce57fa72ced8286f3e74b0b991e33e24a88850
Change-Id: I9b39aab5a61cf12ca1a430deab1076a7a139c692