tests: Avoid loadCallback() in tests and re-enable onMwLoadEnd test
The "onMwLoadEnd" test was flaky due to a previous test calling
loadCallback, which has numerous uncontrolled side-effects such
as from setupFirstInputTimingObserver which sets up an overserver
and then continues to run in the background, this is inherently
unstable to test.

Instead, break it down a little bit into smaller units with
isRegularNavigation() and getOversampleReasons() and test those
directly.

Bug: T299780
Change-Id: I2f57aeb1e7bd716e5115015b7a791be8abd669d5