ext.navigationTiming: Use native packageFiles for virtual config.json
Instead of doing this manually via dynamically generated callback,
use the simpler syntax for exporting config variables.

The dynamic callback can be useful if the information comes from
a different source or needs to be changed, but when it's just exporting
plainly as-is, there is no need for the callback and extra code complexity.

Docs at <https://www.mediawiki.org/wiki/ResourceLoader/Package_files>.

Part of the callback was checking that the $wgNavigationTimingOversampleFactor
contains a number between 0.0 and 1.0 (instead of negative or over 1),
but that seems unnecessary. It is not something we normally do in
MW development, and would be caught soon enough when the change is
reviewed locally, or in Beta Cluster, or during pre-deploy staging.

Bug: T315796
Change-Id: I4980eb493536c115aea56dbdf0f7a7d71d46fa73