Remove ElementTiming measurements
This was something Gilles added to be able to compare the API with
synthetic testing. I think we should collect Element Timings in the
future however we don't need to know size/position etc. Render time
and name should be enough and lets adds it when product has use case
for it.

Also removed references to the Resource Timing API that we don't use.

Bug: T286703
Change-Id: I77a9d757a3de6ce5b283e183604604831d39cd63