tests: Add missing coverage for dnsLookup, netinfo, and more
Replace partial and ad-hoc assertions with a single explicit
assertion for the entire event. For this to work, a few missing
factors needed to be mocked on the input side:

* mw.config wgVersion (added)
* mw.config wgNamespaceNumber (moved up from another test)
* mw.config wgAction (moved up from another test)
* mw.config wgCanonicalSpecialPageName (moved up from another test)
* mw.user.getPageviewToken (moved up from another test)

The rest was already mocked, but even netinfo and dns weren't covered
by tests previously despite being mocked.

This change removes the need for weak type assertions, timestamp range
checks range, and property non-existence checks, as these are all
obsoleted by the single assert.propEqual() check that validates
the whole event object.

Change-Id: Ie1a28db21210be10fc12c489754526beb75d155b