Remove oversample functionality.
Remove the oversample functionality since we don't need that when
we moved to Promethues. This also make the extension leaner and
cleaner. We keep the isOversample and always make it false to make
sure the schema and dashboards work as before.

I also changed the export from emitNavTiming to the full name
emitNavigationTiming so that we have the same standard for most
exports.

Bug: T337858
Change-Id: Ib433f655d154f6e076adc27ee8d9e9f20c93e3c2